pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-google-genai (>=2.1.8,<3.0.0)",
    "streamlit (>=1.47.0,<2.0.0)",
    "pypdf (>=5.8.0,<6.0.0)",
    "faiss-cpu (>=1.11.0.post1,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
